Try to fix the test error of Hudson
git-svn-id: https://svn.apache.org/repos/asf/camel/trunk@773452 13f79535-47bb-0310-9956-ffa450edef68

+++ tests/camel-itest-osgi/src/test/java/org/apache/camel/itest/osgi/OSGiIntegrationSpringTestSupport.java
@@ -55,7 +55,10 @@ public abstract class OSGiIntegrationSpringTestSupport extends OSGiIntegrationTe
     public void tearDown() throws Exception {
         super.tearDown();
         if (applicationContext != null) {
-            applicationContext.destroy();
+            // Try to fix the test error on Hudson
+            if (applicationContext.isActive()) {
+                applicationContext.destroy();
+            }
         }
     }
